Moved Auto-specific classes into separate package, renamed Auto
