+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/autoclasses/GrabberA.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.robot;
+package org.firstinspires.ftc.teamcode.robot.autoclasses;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
